signal fix

+++ site/common/models/mongo/UserSignal.php
@@ -165,7 +165,7 @@ class UserSignal extends EMongoDocument
             $class_name = $this->item_name;
             if (method_exists($class_name::model(), 'getUrl')) {
                 $user = $class_name::model()->findByPk($this->item_id);
-                if ($user === null || $user->deleted){
+                if ($user === null || (isset($user->deleted) && $user->deleted == 1)){
                     $this->delete();
                     return 'error';
                 }
